+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-calendar
 
+## In progress
+
+* Disable 'Actions' menu on DCB Calendar. Refs UICAL-285.
+
 ## [11.0.1] (https://github.com/folio-org/ui-calendar/tree/v11.0.1) (2024-05-08)
 
 * Address empty fourth pane on open of 'Current calendar assignment'. Refs UICAL-283